- re-apply changes   last commit was made by modifying code from an exsHost, however, the   code was not current with the most recent changes in the repo and   as a result when committing a few changes were effectively reverted
Since the code on an esxHost is not under git, this is a present danger
that can be mitigated by having a git clone on the server even though
git will not run there.  If any changes are made then the whole repo
directory can be copied to a machine which has git, and then any commit
and push will be consistent and safe.